// src/app/core/interfaces/paginated-response.model.ts
/**
 * Интерфейс для пагинированного ответа от API.
 * @template T Тип элементов в списке.
 */
export interface PaginatedResponse<T> {
  /**
   * Список элементов в текущей странице.
   */
  items: T[];

  /**
   * Общее количество элементов в полном наборе данных.
   */
  totalCount: number;
}